requirements.txt
appdirs==1.4.4
astroid==3.3.8
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
cssselect==1.2.0
dill==0.3.9
fake-useragent==2.0.3
feedparser==6.0.11
finnhub-python==2.4.22
frozendict==2.4.6
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
lxml==5.3.0
mccabe==0.7.0
multitasking==0.0.11
numpy==2.2.1
pandas==2.2.3
parse==1.20.2
peewee==3.17.8
platformdirs==4.3.6
pyee==11.1.1
pylint==3.3.3
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.17.0
smbus==1.1.post2
soupsieve==2.6
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yahoo-finance==1.4.0
yfinance==0.2.51
zipp==3.21.0
